refactor shared get_list code
