+++ tests/test_get_record.py
@@ -13,6 +13,14 @@ def test_good_params(test_client):
     text=response.get_data(as_text=True)
     assert "<error code='badArgument'>" not in text
 
+def test_bad_meta_format(test_client):
+    params = {OAIParams.VERB: OAIVerbs.GET_RECORD, OAIParams.ID: "oai:arXiv.org:2307.10651",  OAIParams.META_PREFIX: "pictures"}
+    response = test_client.get("/oai", query_string=params)
+    assert response.status_code == 200 
+    text=response.get_data(as_text=True)
+    assert "<error code='cannotDisseminateFormat'>" in text
+    assert "Did not recognize requested format" in text
+
 def test_bad_id(test_client):
 
     params = {OAIParams.VERB: OAIVerbs.GET_RECORD, OAIParams.ID: "arXiv.org:2307.10651",  OAIParams.META_PREFIX: "oai_dc"}
