--- tests/test_list_indentifiers.py
@@ -1,80 +1,0 @@
-from oaipmh.data.oai_properties import OAIParams, OAIVerbs
-
-def test_good_params(test_client):
-    #good minimal params
-    params = {OAIParams.VERB: OAIVerbs.LIST_IDS, OAIParams.META_PREFIX: "oai_dc"}
-    response = test_client.get("/oai", query_string=params)
-    assert response.status_code == 200 
-    text=response.get_data(as_text=True)
-    assert "<error code='badArgument'>" not in text
-
-    response = test_client.post("/oai", data=params)
-    assert response.status_code == 200 
-    text=response.get_data(as_text=True)
-    assert "<error code='badArgument'>" not in text
-
-    #good maximal params
-    params = {OAIParams.VERB: OAIVerbs.LIST_IDS, OAIParams.META_PREFIX: "oai_dc", OAIParams.FROM: "now", OAIParams.UNTIL:"later", OAIParams.SET: "math"}
-    response = test_client.get("/oai", query_string=params)
-    assert response.status_code == 200 
-    text=response.get_data(as_text=True)
-    assert "<error code='badArgument'>" not in text
-
-    response = test_client.post("/oai", data=params)
-    assert response.status_code == 200 
-    text=response.get_data(as_text=True)
-    assert "<error code='badArgument'>" not in text
-
-    #good partial params
-    params = {OAIParams.VERB: OAIVerbs.LIST_IDS, OAIParams.META_PREFIX: "oai_dc", OAIParams.UNTIL:"later", OAIParams.SET: "math"}
-    response = test_client.get("/oai", query_string=params)
-    assert response.status_code == 200 
-    text=response.get_data(as_text=True)
-    assert "<error code='badArgument'>" not in text
-
-    response = test_client.post("/oai", data=params)
-    assert response.status_code == 200 
-    text=response.get_data(as_text=True)
-    assert "<error code='badArgument'>" not in text
-
-def test_extra_params(test_client):
-    #good minimal params
-    params = {OAIParams.VERB: OAIVerbs.LIST_IDS, OAIParams.META_PREFIX: "oai_dc", "color":"green"}
-    response = test_client.get("/oai", query_string=params)
-    assert response.status_code == 200 
-    text=response.get_data(as_text=True)
-    assert "<error code='badArgument'>" in text
-    assert "Unallowed parameter." in text
-
-    response = test_client.post("/oai", data=params)
-    assert response.status_code == 200 
-    text=response.get_data(as_text=True)
-    assert "<error code='badArgument'>" in text
-    assert "Unallowed parameter." in text
-
-def test_token_params(test_client):
-    #correct params
-    params = {OAIParams.VERB: OAIVerbs.LIST_IDS, OAIParams.RES_TOKEN: "rainbow"}
-    response = test_client.get("/oai", query_string=params)
-    assert response.status_code == 200 
-    text=response.get_data(as_text=True)
-    assert "<error code='badArgument'>" not in text
-
-    response = test_client.post("/oai", data=params)
-    assert response.status_code == 200 
-    text=response.get_data(as_text=True)
-    assert "<error code='badArgument'>" not in text
-
-    #cant have other valid params along with token
-    params = {OAIParams.VERB: OAIVerbs.LIST_IDS, OAIParams.RES_TOKEN: "rainbow",  OAIParams.META_PREFIX: "oai_dc"}
-    response = test_client.get("/oai", query_string=params)
-    assert response.status_code == 200 
-    text=response.get_data(as_text=True)
-    assert "<error code='badArgument'>" in text
-    assert "No other paramters allowed with" in text
-
-    response = test_client.post("/oai", data=params)
-    assert response.status_code == 200 
-    text=response.get_data(as_text=True)
-    assert "<error code='badArgument'>" in text
-    assert "No other paramters allowed with" in text
